acme: scrolling fixes + new home/end
Home and End previously navigated between
two different window locations: the top and
the bottom of the text.  Now they include a
third waypoint: the location where typing last
happened.  Thus, in a win window, typing
	ls -l
	<home>
scrolls to the beginning of the ls -l output.
A second <home> continues to the top of the file.

Makes Send scroll always, along with writes by
external programs to +Errors.

R=r
CC=mccoyst
http://codereview.appspot.com/4830051