acme: textshow the result of an Edit.
Otherwise, the Edit can happen off screen.

This is particularly useful when doing find and replace via
Edit .+/<find>/c/<replace>
Without this change, the replace can happen off screen and you need to scroll to it (pressing ← or → suffices but is annoying).
With this change, it scrolls to the result automatically.

Change-Id: I14bf99934e6fcaea2e746799a7fa0004f5499d4c