Update patch set 1
Patch Set 1:

To remove the leading tab was a typo.  I can push the change without it.
About to use $PLAN9/config, it is not included at this point, so it cannot be used. Of course it can be sourced but I think this should be another change, because it is not related to the NetBSD problem.
I have another change in the directory src/cmd/9term/, where it is needed a pty implementation for NetBSD. It still has some problems that I have to fix.

Patch-set: 1