Open files specified on the command line even in the presence of -l.
Before this change, if the -l argument was given, filenames listed on the command line were ignored. With this change, it now loads the dump and then opens the requested files. This is handy for users who have a shell alias that always loads a dump file containing their favorite fonts and column layout.

Change-Id: I3ebec4461d07a3d877f4289712feab46f21297ed