Update patch set 1
Patch Set 1:

> I don't think this is right. As you point out, we don't want this
 > for all commands, and it's very hard to tell which ones you do want
 > it for. The first change? The last change?
 
On the other hand, we are willing to set dot, and therefore the selected text, to the last change. Why not show it? What if the logic was to do nothing if any of dot is already on the screen, but to textshow if not? That way, Edit ,s/this/that will behave the same as without this CL (dot—everything—will be already shown), but other edits that take place off the screen will show?

 > Note that if you do want to find the matches there is always
 > 
 > Edit .+/<find>/=
 > 
 > and then you can poke around before doing the replacement. It's not
 > quite as pretty but you can also do:
 > 
 > Edit ,x/<find>/{c/<replace>/
 > =
 > }

Thanks!

Patch-set: 1