font/fixed: adjust the ascent for the 7x13 subfonts
The ascent in these subfonts was 10 pixels but the images
are using an 11 ascent, so fix all the ascents in the files.
This is the global ascent, not for each character. This change
makes them consistent with the usage in, for instance,
the lucsans directory.

Problem was found with some new Go code to handle these
fonts, and we'd like to publish the updated files so the Go
code can use them in a test. If preferable, of course, they could
be put under some other name.

Others may also want adjustment, but we can start here.

Change-Id: I6c518604eed5b35db641dc537e4b5a2810ad1cf9