mk: Clean up symtab.c, change hash function a bit
syminit is only called once in main(), and at that point the global
array hash is still all zeroes, thus the call is a no-op; it gets
optimized out.
symdel and symstat also aren't used anywhere so they were commented out
and removed from the header.

The hash function was abstracted out of symlook and symdel, and changed
so as not to use signed integer overflow because that is forbidden in
standard C. (-fwrapv can be used to specify that signed overflow wraps
around.)
Also, the hash table size/modulus was increased as it seemed very small.

Change-Id: I5478ae03f208a92a6ba16060486b946dcc4b2cf1