getcallerpc: Return a constant for x86-64
The version of getcallerpc for amd64 was (by mistake?) the exact copy of
the i386 version. Since in the System V AMD64 ABI the first 6 integer
arguments get passed to functions in registers, getcallerpc can't AFAIK
be implemented without GCC's builtin for this architecture.
(See http://www.x86-64.org/documentation/abi.pdf)

As is, the parameter of which the address is taken gets spilled on the
stack, and getcallerpc loads a (possibly even misaligned) wrong value.

Change-Id: I5699c526a37cbceadf759adfa5a077d4156ed815