getcallerpc: Return a constant for x86-64
The version of getcallerpc for amd64 was (by mistake?) the exact copy of
the i386 version. Since in the System V AMD64 ABI the first 6 integer
arguments get passed to functions in registers, getcallerpc can't AFAIK
be implemented without GCC's builtin for this architecture.
(See http://www.x86-64.org/documentation/abi.pdf)

As is, the parameter whose address we take gets spilled on the stack and
getcallerpc loads a (possibly even misaligned) wrong value.

As getcallerpc is only used for debugging the allocation functions, it
seems alright to just return a recognizable constant.

Change-Id: I5699c526a37cbceadf759adfa5a077d4156ed815