Update patch set 1
Patch Set 1:

(3 comments)

This will invalidate old acme.dump files, which is not ok. I understand the need to escape \n but that should be done some other way that leaves (most) existing dump files unaffected. I suggest replacing \n with 0xF000+'\n' before encoding, and then doing the reverse transformation before decoding. That part of the rune space is already taken by keyboard.h, and 0xF000+'\n' is 0xF00A, which is Kprint, which is unlikely to appear in a real tag and can be redefined in this context.

You shouldn't even need helper functions at this point. Just modify the code at the point where it is needed. Should be just a simple loop over the runes in both cases (and another loop to restore the old tag memory after save).

The comments below are good to remember for future code, but they shouldn't matter once the above suggestion happens: there should be no new allocations or functions needed.

Patch-set: 1