Remove WSYSTYPE safeguard
The WSYSTYPE safeguard causes rio to fail to build on macOS & previous versions, both when installing and running 'mk' manually. It provides no verbose, and the user is left to debug on their own. I'm not sure what the original use of the safeguard was, but I can't think of one, and it seems a little outdated.