acme: allow multi-line window tag dump/load encoded with unicode PUA
When acme 'Dump's, multi-line window tags are cut off before newline,
'\n', and the 'acme.dump' only contains the first line of a window
tag.

This commit replaces all '\n' in window tags with a rune (0XF00A) in
unicode PUA region, when 'Dump', and converts them back to '\n' when
'Load'.

In the dump file, due to utf-8 encoding, the replacement rune (0XF00A)
in window tags has the on-disk byte sequence of 0XEF808A.

If you had worked with the rune, 0XF00A, explicitly in window tags,
this commit would replace those in the dump file with '\n' upon
'Load'.