devdraw: rewrite cocoa renderer
The old renderer did a lot of work outside the Cocoa drawRect method.
In macOS 10.14 this stopped working (see
https://github.com/9fans/plan9port/issues/185). The best information
I could find on how to remedy this was the suggestion to "just use
drawRect".

This ended up as a substantial rewrite of the rendering code. It
looks to me like we did not take a performance hit in doing so. I
think the code is simpler now because it lets Cocoa take care of
more things for us.

I have manually tested this code on macOS 10.13 (retina) and 10.14
(non-retina). I don't know how it behaves on 10.12 and before.

To clean up while I'm here I have removed ifdef's in cocoa-screen.m
for 10.6 support. The last 10.6 release was in 2009.