all: add fuzzing and testing support; integrated with oss-fuzz, travis
This pull request sets up fuzzing (a method of finding defects in
software) of plan9port in a way that can be integrated with Google's
OSS-Fuzz project, which will fuzz plan9port with American Fuzzy Lop and
libFuzzer for free. See #281

Currently fuzzing is mostly used to check stuff like buffer overflows,
in the future we could also use it to compare with different APIs that
implement the same standard algorithms, somewhat like what is done in
this commit in the lib9 encoder targets.

The INSTALL script has gained the option "-t" for running regression
tests.

For trying out fuzzing, look at the comment at the beggining of
testing/buildFuzz.sh .

Most of the files added in this commit are test input data (those in
"corpus" directories).

Because of the unsuitability of the existing build system for building
with sanitizers and libFuzzer, a somewhat separate build system was
created that still leverages existing mkfiles for building the plan9port
libraries by hijacking 9c and 9ar from under mk.

Some fuzz targets are linux specific, for example those that use the
memfd_create system call (accessed through libc with the syscall
function for compatibility with libc-s that do not have the memfd_create
function.

Some fuzz targets are disabled for OSS-Fuzz, because they are considered
too small by OSS-Fuzz Googlers.

Interesting links:

https://github.com/google/oss-fuzz/tree/master/projects/example/

https://google.github.io/oss-fuzz/

https://llvm.org/docs/LibFuzzer.html

http://lcamtuf.coredump.cx/afl/

https://github.com/google/AFL

Updates #281

Change-Id: I458f00ebbc1bb39b400fcef5e350a59b36f4923d