cmd/yacc: refactor to enable fuzzing
This changes yacc to only exit directly from (p9)main; and exports three
formerly function-scope static variables, so we could reinitialize them
in the fuzzer (this is important for an in-process fuzzer).

This is necessary for fuzzing. By changing the yacc internal functions
to report errors instead of exiting, we can replace main and fuzz them.

Change-Id: I8fef0595040160c621f9e1e98ab4efb6db14e10e