libthread: remove adjustments to stack pointer/size
When setting the stack pointer and size in the
ucontext structure when creating a thread, the
stack pointer was offset by 8 bytes and the size
was 64 less than allocated.  This appears to be
debugging code from some time ago, but I couldn't
quite figure out what it's purpose was (perhaps
poisoning the top/bottom of the stack to detect
under/overflow?).  The 8-byte offset for the
stack pointer, in particular, was causing instructions
like `MOVAPS` to fail on some platforms: the ABI
assums 16-byte stack alignment on entry to a
function.

Anyway, remove the adjustments.

Fixes #109

Signed-off-by: Dan Cross <cross@gajendra.net>