rc: implement 9atoms `sep{command} syntax
After a bunch of experience using this, I
find that it  makes a good deal of code
far simpler to write. For example,

	nl='
	'

	for(f in `$nl{find . -type f})
		munge $f

ends up quite a bit more pleasant to use
for handling file output line by line,
since it doesn't cause scoping issues
for rc functions, and is more explicit
about what is being split on.