acme: add x11 mappings
The macOS bindings are present, but the X11 ones are not. Redo is mapped
to ^Y because the shift modifier is not translated.

This patch works, but I'd prefer if we implemented a modifier system
over keys. For example, we remove `Kctl`, `Kcmd`, `Kshift` and `Kalt`
and instead make these "modifier" keys:

```c
enum {
    Mctl= 0x0100,
    Malt=   0x0200,
    Mshift= 0x0400,
    Msuper= 0x0800,
};
```

This way, the switch can be rewritten to be:

```c
case Mctl|Mshift+'z': /* shift-^Z: redo */
```

Please let me know if this is a reasonable approach to this issue and
something you'd prefer, or if this change is adequate.