rio: pass b2 and b3 to clients
The existing call to XAllowClients to ReplayPointer is ineffective
because the pointer is not frozen at the time.

To make it effective, we change XGrabTutton to grab the pointer
synchronously. In order not to freeze the window manager though, we
need to always call XAllowEvents after processing a mouse event. We
use ReplayPointer if the button isn't b1.

References:
* https://tronche.com/gui/x/xlib/input/XAllowEvents.html
* https://tronche.com/gui/x/xlib/input/XGrabButton.html
* https://tronche.com/gui/x/xlib/input/XGrabPointer.html