Replace stunnel3 with openssl in imap and smtp dial, change smtp port to 465
`stunnel3` is more than 20 years old years already, and it is not
present in most operating systems, there are new versions of it, but
work differently than the old one. This patch replace it in the mail
commands with openssl, which I believe is more common.

Lastly, change the default smtp port from 587 to 465 as per RFC 8413
recommendation.

https://datatracker.ietf.org/doc/html/rfc8314#section-3