INSTALL: Use GNU AWK on AIX
When running the install script on AIX, the build pipes most of the
output through dist/isum.awk.  This AWK script does not display any
output on AIX using the default awk located in /usr/bin/awk.  On
the other platforms, this script displays the current item being
compiled on a single line and updates that line as the build
progresses.  Using GNU AWK, which is available from the IBM AIX
Toolbox for Open Source Software, allows the script to run
correctly.  This makes for a prettier build at the cost of
requiring gawk on AIX.  During the run of INSTALL,
/opt/freeware/bin must also be in the PATH environment variable.