INSTALL: Detect fontsrv X11 prerequsites on AIX
The INSTALL script detects if the fontsrv prerequsites are installed
by compiling a test program.  However, this does not work on AIX
even with freetype installed from the IBM Toolbox for Open Source
Software.  This does not work for two reasons.  First, the IBM-
supplied package is installed into /opt/freeware, and therefore
we need to include the following directories:
/opt/freeware/include
/opt/freeware/include/freetype2

Secondly, the INSTALL script uses cc to compile the script, and
this does not set up enough PLAN9 environment for the compiler
to find the right headers on AIX.  Therefore, we use 9c which
handles all of the AIX compiler envionment setup, and should also
work for the other platforms.