INSTALL: Log compiler version during AIX build
The INSTALL script typically prints the version of the compiler in use
during a build, and does so by running "9c -v".  Ultimately the -v
argument gets passed to the actual compiler.  However, the AIX compiler
(XL C) does not use -v to display the compiler version, but instead
uses the argument -qversion.  Therefore, we add -qversion to the 9c
invocation on this platform only.