fontsrv: Fix build on AIX
The fonsrv program can build on AIX so long as the prerequsites are
installed.  These include freetype2 and fontconfig, both of which
are available from the IBM Toolbox for Open Source Software.
However, these get installed into /opt/freeware.  Update the
freetyperules.sh script to look for the headers and libraries in
/opt/freeware on AIX.