Don't call gfx_abortcompose with a lock held.
There has been some funny locking/stalling behavior that seems to be related to using alt+click as a simulated middle-click. It may have started after this gfx_abortcompose call was added. According to devdraw.h, the gfx_ routines should not be called with a lock held. Now I'm not sure this fixes the weird behavior I've seen, but can't hurt to still fix this buggy call.